moving add-category modal to separate file
